fix(rust, python): true-divide output type
