+++ polars/polars-lazy/polars-plan/src/logical_plan/aexpr/schema.rs
@@ -1,185 +1,6 @@
-use std::sync::Arc;
-
-use polars_arrow::prelude::QuantileInterpolOptions;
-use polars_core::prelude::*;
-use polars_core::utils::{get_time_units, try_get_supertype};
-use polars_utils::arena::{Arena, Node};
-
-use crate::dsl::function_expr::FunctionExpr;
-use crate::logical_plan::Context;
-use crate::prelude::names::COUNT;
-use crate::prelude::*;
-
-#[derive(Clone, Debug)]
-pub enum AAggExpr {
-    Min {
-        input: Node,
-        propagate_nans: bool,
-    },
-    Max {
-        input: Node,
-        propagate_nans: bool,
-    },
-    Median(Node),
-    NUnique(Node),
-    First(Node),
-    Last(Node),
-    Mean(Node),
-    List(Node),
-    Quantile {
-        expr: Node,
-        quantile: Node,
-        interpol: QuantileInterpolOptions,
-    },
-    Sum(Node),
-    Count(Node),
-    Std(Node, u8),
-    Var(Node, u8),
-    AggGroups(Node),
-}
-
-// AExpr representation of Nodes which are allocated in an Arena
-#[derive(Clone, Debug, Default)]
-pub enum AExpr {
-    Explode(Node),
-    Alias(Node, Arc<str>),
-    Column(Arc<str>),
-    Literal(LiteralValue),
-    BinaryExpr {
-        left: Node,
-        op: Operator,
-        right: Node,
-    },
-    Cast {
-        expr: Node,
-        data_type: DataType,
-        strict: bool,
-    },
-    Sort {
-        expr: Node,
-        options: SortOptions,
-    },
-    Take {
-        expr: Node,
-        idx: Node,
-    },
-    SortBy {
-        expr: Node,
-        by: Vec<Node>,
-        reverse: Vec<bool>,
-    },
-    Filter {
-        input: Node,
-        by: Node,
-    },
-    Agg(AAggExpr),
-    Ternary {
-        predicate: Node,
-        truthy: Node,
-        falsy: Node,
-    },
-    AnonymousFunction {
-        input: Vec<Node>,
-        function: SpecialEq<Arc<dyn SeriesUdf>>,
-        output_type: GetOutput,
-        options: FunctionOptions,
-    },
-    Function {
-        /// function arguments
-        input: Vec<Node>,
-        /// function to apply
-        function: FunctionExpr,
-        options: FunctionOptions,
-    },
-    Window {
-        function: Node,
-        partition_by: Vec<Node>,
-        order_by: Option<Node>,
-        options: WindowOptions,
-    },
-    #[default]
-    Wildcard,
-    Slice {
-        input: Node,
-        offset: Node,
-        length: Node,
-    },
-    Count,
-    Nth(i64),
-}
+use super::*;
 
 impl AExpr {
-    /// Any expression that is sensitive to the number of elements in a group
-    /// - Aggregations
-    /// - Sorts
-    /// - Counts
-    /// - ..
-    pub(crate) fn groups_sensitive(&self) -> bool {
-        use AExpr::*;
-        match self {
-            Function { options, .. } | AnonymousFunction { options, .. } => {
-                options.is_groups_sensitive()
-            }
-            Sort { .. }
-            | SortBy { .. }
-            | Agg { .. }
-            | Window { .. }
-            | Count
-            | Slice { .. }
-            | Take { .. }
-            | Nth(_)
-             => true,
-            | Alias(_, _)
-            | Explode(_)
-            | Column(_)
-            | Literal(_)
-            // a caller should traverse binary and ternary
-            // to determine if the whole expr. is group sensitive
-            | BinaryExpr { .. }
-            | Ternary { .. }
-            | Wildcard
-            | Cast { .. }
-            | Filter { .. } => false,
-        }
-    }
-
-    /// This should be a 1 on 1 copy of the get_type method of Expr until Expr is completely phased out.
-    pub fn get_type(
-        &self,
-        schema: &Schema,
-        ctxt: Context,
-        arena: &Arena<AExpr>,
-    ) -> PolarsResult<DataType> {
-        self.to_field(schema, ctxt, arena)
-            .map(|f| f.data_type().clone())
-    }
-
-    pub(crate) fn replace_input(self, input: Node) -> Self {
-        use AExpr::*;
-        match self {
-            Alias(_, name) => Alias(input, name),
-            Cast {
-                expr: _,
-                data_type,
-                strict,
-            } => Cast {
-                expr: input,
-                data_type,
-                strict,
-            },
-            _ => todo!(),
-        }
-    }
-
-    pub(crate) fn get_input(&self) -> Node {
-        use AExpr::*;
-        match self {
-            Alias(input, _) => *input,
-            Cast { expr, .. } => *expr,
-            _ => todo!(),
-        }
-    }
-
     /// Get Field result of the expression. The schema is the input data.
     pub fn to_field(
         &self,
@@ -240,32 +61,10 @@ impl AExpr {
                             out_field = arena.get(*left).to_field(schema, ctxt, arena)?;
                             out_field.name().as_str()
                         };
-                        Field::new(out_name, DataType::Boolean)
+                        Field::new(out_name, Boolean)
                     }
-                    _ => {
-                        // don't traverse tree until strictly needed. Can have terrible performance.
-                        // # 3210
-
-                        // take the left field as a whole.
-                        // don't take dtype and name separate as that splits the tree every node
-                        // leading to quadratic behavior. # 4736
-                        let mut left_field = arena.get(*left).to_field(schema, ctxt, arena)?;
-                        let right_type = arena.get(*right).get_type(schema, ctxt, arena)?;
-
-                        let super_type = match op {
-                            Operator::Minus => match (&left_field.dtype, right_type) {
-                                // T - T != T if T is a datetime / date
-                                (Datetime(tul, _), Datetime(tur, _)) => {
-                                    Duration(get_time_units(tul, &tur))
-                                }
-                                (Date, Date) => Duration(TimeUnit::Milliseconds),
-                                (left, right) => try_get_supertype(left, &right)?,
-                            },
-                            _ => try_get_supertype(&left_field.dtype, &right_type)?,
-                        };
-                        left_field.coerce(super_type);
-                        left_field
-                    }
+                    Operator::TrueDivide => return get_truediv_field(*left, arena, ctxt, schema),
+                    _ => return get_arithmetic_field(*left, *right, arena, *op, ctxt, schema),
                 };
 
                 Ok(field)
@@ -395,8 +194,58 @@ impl AExpr {
     }
 }
 
+fn get_arithmetic_field(
+    left: Node,
+    right: Node,
+    arena: &Arena<AExpr>,
+    op: Operator,
+    ctxt: Context,
+    schema: &Schema,
+) -> PolarsResult<Field> {
+    // don't traverse tree until strictly needed. Can have terrible performance.
+    // # 3210
+
+    // take the left field as a whole.
+    // don't take dtype and name separate as that splits the tree every node
+    // leading to quadratic behavior. # 4736
+    use DataType::*;
+    let mut left_field = arena.get(left).to_field(schema, ctxt, arena)?;
+    let right_type = arena.get(right).get_type(schema, ctxt, arena)?;
+
+    let super_type = match op {
+        Operator::Minus => match (&left_field.dtype, right_type) {
+            // T - T != T if T is a datetime / date
+            (Datetime(tul, _), Datetime(tur, _)) => Duration(get_time_units(tul, &tur)),
+            (Date, Date) => Duration(TimeUnit::Milliseconds),
+            (left, right) => try_get_supertype(left, &right)?,
+        },
+        _ => try_get_supertype(&left_field.dtype, &right_type)?,
+    };
+    left_field.coerce(super_type);
+    Ok(left_field)
+}
+
 fn coerce_numeric_aggregation(field: &mut Field) {
     if field.dtype.is_numeric() && !matches!(&field.dtype, DataType::Float32) {
         field.coerce(DataType::Float64)
     }
+}
+
+fn get_truediv_field(
+    left: Node,
+    arena: &Arena<AExpr>,
+    ctxt: Context,
+    schema: &Schema,
+) -> PolarsResult<Field> {
+    let mut left_field = arena.get(left).to_field(schema, ctxt, arena)?;
+    use DataType::*;
+    let out_type = match left_field.data_type() {
+        Float32 => Float32,
+        dt if dt.is_numeric() => Float64,
+        // we don't know what to do here, best return the dtype
+        dt => dt.clone(),
+    };
+
+    left_field.coerce(out_type);
+    Ok(left_field)
 }
